[상점] 액션 개발 완료. 상점의 판매 목록, 상점에 판 목록 UI

+++ nekoyume/Assets/_Scripts/Action/Context.cs
@@ -10,7 +10,7 @@ namespace Nekoyume.Action
     {
         public Avatar avatar;
         public BattleLog battleLog;
-        public int gold;
+        public decimal gold;
         public DateTimeOffset updatedAt;
         public DateTimeOffset? clearedAt;
         public Address? AvatarAddress { get; private set; }
